chore: run oxlint on all files during lint-staged (for now)
Currently when a specific file is passed to oxlint (e.g. `/pwd/to/test.js`),
it will lint the file despite having it specified in `.eslintignore`,
this makes lint-staged useless when an ignored file is committed.

oxlint has the same behavior as eslint.

This PR makes lint-staged to run on all changed files.
The overhead is manageable because oxlint completes within a second.

See
* https://github.com/okonet/lint-staged#how-can-i-ignore-files-from-eslintignore
* https://github.com/okonet/lint-staged#example-run-tsc-on-changes-to-typescript-files-but-do-not-pass-any-filename-arguments