dev: compatible with vanilla extract style (#4932)
* feat: add vanilla extract css example

* feat: add `examples/vanilla-extract-css`

* feat(externals-plugin): filter falsy external entries before processing

This commit updates the `ExternalsPlugin` to filter out any falsy values in the provided externals array before mapping them to raw external items. Additionally, it enhances the type signature of `getRawExternalItem` to accept an undefined input, and adds handling for unexpected item types by throwing a `TypeError`.

The introduced changes prevent potential runtime errors that can occur when undefined or otherwise falsy values are inadvertently passed into the `externals` array, ensuring the build process is more robust and predictable. The added error handling provides a clearer diagnostic message when an invalid configuration is encountered, which will aid in debugging and maintaining the codebase.

* feat(rspack): export EntryOptionPlugin from rspack package

This commit adds the export of `EntryOptionPlugin` to the public API of the rspack package, allowing users to import it directly. The new export is added above the Internal Plugins section, indicating that it is intended for public use.

* refactor(rspack): comment out createModule hook and associated types

The commit comments out unused code related to the `CreateData` type and the `createModule` hook within the `NormalModuleFactory.ts`. This change suggests a cleanup for better readability or preparation for future refactoring. Future development may leverage these hooks and types, so they have been commented rather than deleted.

The addition of these comments should not affect the current functionality of rspack, but developers should be aware that references to `CreateData` and `createModule` are currently inactive.

* chore: update pnpm lock file

* test: add repro

* feat: parse resource wihout fragment for loader context

* feat: update rspack config

* feat: implement basic normal_module_factory_create_module hook

* chore: remove useless files from vanilla extract css example

* feat: update error handling

* feat: update vanilla extract variables

* feat: add stub args to nmf createModule

* test: add test case for createModule hook

* fix: failed to get CreateModuleData from NormalModuleCreateData

* feat: add matchResource for params of normalModuleFactory.hooks.matchResource

* feat: fix for passing clippy