fix: Use "latest" tag for canary release (#10538)
I have noticed that in the documentation its suggested to use `latest` tag for the rspack-canary packages. But latest released version for the package actually has a `canary` tag. The version by `latest` tag seems to be outdated and not updated anymore.

Because of the alleged inconsistency of documentation and versions tags, I suggest changing the tag in the canary release workflow from `canary` to `latest`.

Alternative would be to update the documentation.

Links:
https://rspack.rs/guide/start/quick-start#install-canary-version

https://www.npmjs.com/package/@rspack-canary/core?activeTab=versions