fix: Narrow the Dev Server Middleware type (#10522)
* fix: Narrow the Dev Server Middleware type

It will be convenient to have the Dev Server Middleware types exported from
the rspack. All our middleware is written in separate modules and annotated
with JSDoc types.
By exporting the type, we can ge rid of webpack-dev-server mentions in the
annotation and use `import('@rspack/core').DevServerMiddleware` to type
the middleware we have.

I have also made an effort to narrow the Middleware type so it is aligned
with the one in the webpack-dev-middleware.

(see https://github.com/webpack/webpack-dev-middleware/blob/master/types/index.d.ts#L370)

* fix: Narrow the Dev Server Middleware type